requirements.txt
numpy==1.23.5
numpy==1.19.5
numpy==1.23.1
opencv_contrib_python==4.6.0.66
opencv_contrib_python_headless==4.7.0.72
opencv_python==4.6.0.66
opencv_python_headless==4.7.0.72
Pillow==9.4.0
Pillow==9.5.0
Pillow==10.1.0
streamlit==1.27.2
~pencv_python_headless==4.5.4.60
